check project <-> task relationship in project task routes

+++ app-backend/api/src/main/scala/annotation-project/AnnotationProjectTaskRoutes.scala
@@ -222,21 +222,26 @@ trait AnnotationProjectTaskRoutes
         ScopedAction(Domain.AnnotationProjects, Action.ReadTasks, None),
         user
       ) {
-        {
-          authorizeAuthResultAsync {
+        authorizeAuthResultAsync {
+          (
             AnnotationProjectDao
               .authorized(
                 user,
                 ObjectType.AnnotationProject,
                 projectId,
                 ActionType.Annotate
-              )
-              .transact(xa)
-              .unsafeToFuture
-          } {
-            complete {
-              TaskDao.getTaskWithActions(taskId).transact(xa).unsafeToFuture
-            }
+              ),
+            TaskDao.getTaskById(taskId)
+          ).mapN({
+            case (success @ AuthSuccess(_), Some(task)) =>
+              if (task.annotationProjectId == projectId) success
+              else AuthFailure[AnnotationProject]()
+            case _ => AuthFailure[AnnotationProject]()
+          }).transact(xa)
+            .unsafeToFuture
+        } {
+          complete {
+            TaskDao.getTaskWithActions(taskId).transact(xa).unsafeToFuture
           }
         }
       }
@@ -250,16 +255,20 @@ trait AnnotationProjectTaskRoutes
       ) {
         authorizeAsync {
           (for {
-            auth1 <- AnnotationProjectDao
-              .authorized(
-                user,
-                ObjectType.AnnotationProject,
-                projectId,
-                ActionType.Annotate
-              )
+            auth1 <-
+              AnnotationProjectDao
+                .authorized(
+                  user,
+                  ObjectType.AnnotationProject,
+                  projectId,
+                  ActionType.Annotate
+                )
             auth2 <- TaskDao.isLockingUserOrUnlocked(taskId, user)
+            auth3 <- TaskDao.getTaskById(taskId) map { taskO =>
+              taskO map { _.annotationProjectId }
+            }
           } yield {
-            auth1.toBoolean && auth2
+            auth1.toBoolean && auth2 && auth3 == Some(projectId)
           }).transact(xa).unsafeToFuture
         } {
           entity(as[Task.TaskFeatureCreate]) { tfc =>
@@ -283,14 +292,21 @@ trait AnnotationProjectTaskRoutes
         user
       ) {
         authorizeAuthResultAsync {
-          AnnotationProjectDao
-            .authorized(
-              user,
-              ObjectType.AnnotationProject,
-              projectId,
-              ActionType.Edit
-            )
-            .transact(xa)
+          (
+            AnnotationProjectDao
+              .authorized(
+                user,
+                ObjectType.AnnotationProject,
+                projectId,
+                ActionType.Edit
+              ),
+            TaskDao.getTaskById(taskId)
+          ).mapN({
+            case (success @ AuthSuccess(_), Some(task)) =>
+              if (task.annotationProjectId == projectId) success
+              else AuthFailure[AnnotationProject]()
+            case _ => AuthFailure[AnnotationProject]()
+          }).transact(xa)
             .unsafeToFuture
         } {
           complete {
@@ -301,13 +317,14 @@ trait AnnotationProjectTaskRoutes
     }
 
   def lockTask(projectId: UUID, taskId: UUID): Route =
-    toggleLock(projectId, TaskDao.lockTask(taskId))
+    toggleLock(projectId, taskId, TaskDao.lockTask(taskId))
 
   def unlockTask(projectId: UUID, taskId: UUID): Route =
-    toggleLock(projectId, _ => TaskDao.unlockTask(taskId))
+    toggleLock(projectId, taskId, _ => TaskDao.unlockTask(taskId))
 
   private def toggleLock(
       projectId: UUID,
+      taskId: UUID,
       f: (User => ConnectionIO[Option[Task.TaskFeature]])
   ): Route =
     authenticate { user =>
@@ -316,13 +333,22 @@ trait AnnotationProjectTaskRoutes
         user
       ) {
         authorizeAsync {
-          AnnotationProjectDao
-            .authorized(
-              user,
-              ObjectType.AnnotationProject,
-              projectId,
-              ActionType.Annotate
-            )
+          (
+            AnnotationProjectDao
+              .authorized(
+                user,
+                ObjectType.AnnotationProject,
+                projectId,
+                ActionType.Annotate
+              ),
+            TaskDao.getTaskById(taskId)
+          )
+            .mapN({
+              case (success @ AuthSuccess(_), Some(task)) =>
+                if (task.annotationProjectId == projectId) success
+                else AuthFailure[AnnotationProject]()
+              case _ => AuthFailure[AnnotationProject]()
+            })
             .transact(xa)
             .map(_.toBoolean)
             .unsafeToFuture
@@ -413,19 +439,23 @@ trait AnnotationProjectTaskRoutes
       ) {
         authorizeAsync {
           (for {
-            auth1 <- AnnotationProjectDao
-              .authorized(
-                user,
-                ObjectType.AnnotationProject,
-                projectId,
-                actionType
-              )
+            auth1 <-
+              AnnotationProjectDao
+                .authorized(
+                  user,
+                  ObjectType.AnnotationProject,
+                  projectId,
+                  actionType
+                )
             auth2 <- TaskDao.hasStatus(
               taskId,
               requiredStatuses
             )
+            auth3 <- TaskDao.getTaskById(taskId) map { taskO =>
+              taskO map { _.annotationProjectId }
+            }
           } yield {
-            auth1.toBoolean && auth2
+            auth1.toBoolean && auth2 && auth3 == Some(projectId)
           }).transact(xa).unsafeToFuture
         } {
           entity(as[AnnotationLabelWithClassesFeatureCollectionCreate]) { fc =>
@@ -434,17 +464,19 @@ trait AnnotationProjectTaskRoutes
             }
             onSuccess(
               (for {
-                _ <- if (deleteBeforeAdding) {
+                _ <-
+                  if (deleteBeforeAdding) {
+                    AnnotationLabelDao
+                      .deleteByProjectIdAndTaskId(projectId, taskId)
+                  } else { 0.pure[ConnectionIO] }
+                insert <-
                   AnnotationLabelDao
-                    .deleteByProjectIdAndTaskId(projectId, taskId)
-                } else { 0.pure[ConnectionIO] }
-                insert <- AnnotationLabelDao
-                  .insertAnnotations(
-                    projectId,
-                    taskId,
-                    annotationLabelWithClassesCreate.toList,
-                    user
-                  )
+                    .insertAnnotations(
+                      projectId,
+                      taskId,
+                      annotationLabelWithClassesCreate.toList,
+                      user
+                    )
               } yield {
                 insert
               }).transact(xa)
@@ -500,13 +532,22 @@ trait AnnotationProjectTaskRoutes
       ) {
         {
           authorizeAuthResultAsync {
-            AnnotationProjectDao
-              .authorized(
-                user,
-                ObjectType.AnnotationProject,
-                projectId,
-                ActionType.Annotate
-              )
+            (
+              AnnotationProjectDao
+                .authorized(
+                  user,
+                  ObjectType.AnnotationProject,
+                  projectId,
+                  ActionType.Annotate
+                ),
+              TaskDao.getTaskById(taskId)
+            )
+              .mapN({
+                case (success @ AuthSuccess(_), Some(task)) =>
+                  if (task.annotationProjectId == projectId) success
+                  else AuthFailure[AnnotationProject]()
+                case _ => AuthFailure[AnnotationProject]()
+              })
               .transact(xa)
               .unsafeToFuture
           } {
@@ -527,14 +568,21 @@ trait AnnotationProjectTaskRoutes
         user
       ) {
         authorizeAuthResultAsync {
-          AnnotationProjectDao
-            .authorized(
-              user,
-              ObjectType.AnnotationProject,
-              projectId,
-              ActionType.Annotate
-            )
-            .transact(xa)
+          (
+            AnnotationProjectDao
+              .authorized(
+                user,
+                ObjectType.AnnotationProject,
+                projectId,
+                ActionType.Annotate
+              ),
+            TaskDao.getTaskById(taskId)
+          ).mapN({
+            case (success @ AuthSuccess(_), Some(task)) =>
+              if (task.annotationProjectId == projectId) success
+              else AuthFailure[AnnotationProject]()
+            case _ => AuthFailure[AnnotationProject]()
+          }).transact(xa)
             .unsafeToFuture
         } {
           complete {
